Create VMTranslator
Because of the finicky online grader, the file creation process here is completely unusable outside of the 6 files that were required to be translated.  In an earlier version, however, it worked perfectly in my local environment and could be used to translate any file, but the grader would not accept it because of file name issues.  I will upload the better version once I finish implementing the staticsTest handler.